chore: update uploading task

+++ vlr/data/tasks/upload.py
@@ -28,28 +28,29 @@ def parse_args():
         help="Path to file or directory to upload.",
     )
     argparser.add_argument(
-        "--channels",
+        "--channel-names-path",
         type=str,
         help="Path to file containing channel names to upload.",
     )
     argparser.add_argument(
-        "--auto-zip",
-        type=bool,
+        "--zip",
+        action=argparse.BooleanOptionalAction,
         default=True,
         help="Whether to automatically zip the directory.",
     )
     argparser.add_argument(
         "--overwrite",
-        type=bool,
+        action=argparse.BooleanOptionalAction,
         default=False,
         help="Whether to overwrite existing zip files.",
     )
     argparser.add_argument(
         "--clean-up",
-        type=bool,
+        action=argparse.BooleanOptionalAction,
         default=False,
         help="Whether to clean up zip files after uploading.",
     )
+    return argparser.parse_args()
 
 
 def get_relative_data_dirs(src_dir: str):
@@ -59,17 +60,11 @@ def get_relative_data_dirs(src_dir: str):
     :return:        Relative paths to data directories.
     """
     relative_data_dirs = []
-    for stage_name in os.listdir(src_dir):
-        relative_data_dir = stage_name
-        stage_dir = os.path.join(src_dir, stage_name)
-        if not os.path.isdir(stage_dir):
+    for data_name in os.listdir(src_dir):
+        data_dir = os.path.join(src_dir, data_name)
+        if not os.path.isdir(data_dir):
             continue
-        for data_name in os.listdir(stage_dir):
-            relative_data_dir = os.path.join(relative_data_dir, data_name)
-            data_dir = os.path.join(stage_dir, data_name)
-            if not os.path.isdir(data_dir):
-                continue
-            relative_data_dirs.append(relative_data_dir)
+        relative_data_dirs.append(data_name)
     return relative_data_dirs
 
 
@@ -77,7 +72,7 @@ def main(args: argparse.Namespace):
     """
     Main function.
     """
-    with open(args.channels, "r") as f:
+    with open(args.channel_names_path, "r") as f:
         channel_names = f.read().splitlines()
 
     relative_data_dirs = get_relative_data_dirs(args.src)
@@ -91,25 +86,37 @@ def main(args: argparse.Namespace):
     ):
         for relative_data_dir in tqdm(
             relative_data_dirs,
-            desc="Processing data directories",
+            desc=f"Processing {channel_name}",
             total=len(relative_data_dirs),
             unit="directory",
             leave=False,
         ):
             channel_dir = os.path.join(args.src, relative_data_dir, channel_name)
-            if not os.path.exists(channel_dir):
-                print(f"Channel {channel_name} does not exist in {relative_data_dir}.")
-                continue
+            if not relative_data_dir.startswith("stage"):
+                if not os.path.exists(channel_dir):
+                    print(f"Channel {channel_name} does not exist in {relative_data_dir}.")
+                    continue
 
-            if args.auto_zip:
-                zip_dir(channel_dir, overwrite=args.overwrite)
+                file_path = channel_dir + ".zip"
+                if args.zip or not os.path.exists(file_path):
+                    zip_dir(channel_dir, overwrite=args.overwrite)
 
-            file_path = channel_dir + ".zip"
+                path_in_repo = os.path.join(
+                    os.path.basename(args.src), relative_data_dir, channel_name + ".zip"
+                )
+            else:
+                file_path = channel_dir + ".json"
+                if not os.path.exists(file_path):
+                    print(f"Channel {file_path} does not exist in {relative_data_dir}.")
+                    continue
+
+                path_in_repo = os.path.join(
+                    os.path.basename(args.src), relative_data_dir, channel_name + ".json"
+                )
+
             api.upload_file(
                 path_or_fileobj=file_path,
-                path_in_repo=os.path.join(
-                    os.path.basename(args.src), relative_data_dir, channel_name + ".zip"
-                ),
+                path_in_repo=path_in_repo,
                 repo_id="fptu/vlr",
                 repo_type="dataset",
                 commit_message=f"chore: update {os.path.basename(relative_data_dir)} directory",
@@ -117,10 +124,9 @@ def main(args: argparse.Namespace):
                 token=args.token,
             )
 
-            if args.clean_up:
+            if args.clean_up and file_path.endswith("zip"):
                 os.remove(file_path)
 
 
 if __name__ == "__main__":
-    args = parse_args()
-    main(args)
+    main(parse_args())
